Moved k6 to load.  Broke out e2e test utils
