+++ integration/load/stress_test_write_path.js
